Refactor Cards to CardList and Card files
